+++ node_helper.js
@@ -5,15 +5,18 @@
  * MIT Licensed.
  */
 
-const NodeHelper = require("node_helper");
-const Log = require("logger");
-const providers = require("./providers");
+/* eslint-env node */
+
+const NodeHelper = require('node_helper');
+const Log = require('logger');
+
+const providers = require('./providers');
 
 module.exports = NodeHelper.create({
-    requiresVersion: "2.15.0",
+    requiresVersion: '2.15.0',
 
     socketNotificationReceived(notification, payload) {
-        if (notification === "CONFIG") {
+        if (notification === 'CONFIG') {
             this.config = payload;
             this.getCountryMedals();
             setInterval(() => {
@@ -27,16 +30,14 @@ module.exports = NodeHelper.create({
             const provider = providers[this.config.provider];
 
             if (!provider) {
-                throw new Error(
-                    `Unsupported provider: ${this.config.provider}`
-                );
+                throw new Error(`Unsupported provider: ${this.config.provider}`);
             }
 
             const countries = await provider.getCountryMedals();
 
-            this.sendSocketNotification("COUNTRIES", countries);
+            this.sendSocketNotification('COUNTRIES', countries);
         } catch (e) {
-            Log.error("Error getting olympic game medals", e);
+            Log.error('Error getting olympic game medals', e);
         }
     }
 });