+++ providers/index.js
@@ -1,4 +1,4 @@
-const bloomberg = require("./bloomberg");
-const nbc = require("./nbc");
+const bloomberg = require('./bloomberg');
+const nbc = require('./nbc');
 
 module.exports = { bloomberg, nbc };