+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 
+* Added option for user to input list of countries to display, see config option `countryList`.
+
 ### Changed
 
 ### Removed