+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # MMM-OlympicGames Changelog
 
+## [1.3.0]
+
+### Added
+
+* Added option for user to input list of countries to display, see config option `countryList`.
+* Added French language translations.
+
+### Changed
+
+* Uniform spelling for MagicMirror²
+* Upgraded development dependencies
+* Updated preview images
+
 ## [1.2.0]
 
 ### Added